requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
ldap3==2.9.1
python-dotenv==1.0.0
Werkzeug==3.0.1
redis==5.0.1
gunicorn==21.2.0
cryptography==41.0.7

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0
